+++ rollup.config.js
@@ -1,14 +1,18 @@
 import babel from 'rollup-plugin-babel';
 import resolve from 'rollup-plugin-node-resolve';
+import pkg from './package.json';
 
 export default {
   input: 'src/index.js',
-  output: {
-    file: 'dist/index.js',
+  output: [{
     format: 'umd',
     name: 'BigTest.Mocha',
-    globals: { mocha: 'mocha' }
-  },
+    globals: { mocha: 'mocha' },
+    file: pkg.main
+  }, {
+    format: 'es',
+    file: pkg.module
+  }],
   external: ['mocha'],
   plugins: [
     resolve(),
@@ -16,8 +20,9 @@ export default {
       babelrc: false,
       comments: false,
       presets: [
-        ['@babel/preset-env', {
-          modules: false
+        ['@babel/env', {
+          modules: false,
+          useBuiltIns: 'usage'
         }]
       ]
     })